% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignificanceTestStuartMaxwell.R
\name{stuartmaxwell_pairwise}
\alias{stuartmaxwell_pairwise}
\title{Stuart Maxwell Test Statistic
Objective :
  To compute pairwise StuartMaxwell statistics for a set of predictions
  1. Find all *.predictions files in the given prediction folder
  2. Create a pairwise set for all prediction files
  3. Load prediction file in pairwise manner
    a. Each prediction file expected to have format label,prediction as columns
  4. Create agreement table based on the 2 sets of predictions
  5. Run Stuart Maxwell on the agreement table
  6. Return Test Statistic, Degree Of Freedom and pValue based on Chi-Square Distribution
  7. Optional: Save results to file, named in pairwise manner.
Usage:
  1. stuartmaxwell_pairwise(predictionFolder = 'Full folder path')
  2. Results will be saved in *.sm and *.smpvalue files
Created by: kah siong
Created on: 05/08/17}
\usage{
stuartmaxwell_pairwise(predictionFolder)
}
\arguments{
\item{A}{full path to the folder containing all the prediction files}
}
\value{
Print out and saved files
}
\description{
Stuart Maxwell Test Statistic
Objective :
  To compute pairwise StuartMaxwell statistics for a set of predictions
  1. Find all *.predictions files in the given prediction folder
  2. Create a pairwise set for all prediction files
  3. Load prediction file in pairwise manner
    a. Each prediction file expected to have format label,prediction as columns
  4. Create agreement table based on the 2 sets of predictions
  5. Run Stuart Maxwell on the agreement table
  6. Return Test Statistic, Degree Of Freedom and pValue based on Chi-Square Distribution
  7. Optional: Save results to file, named in pairwise manner.
Usage:
  1. stuartmaxwell_pairwise(predictionFolder = 'Full folder path')
  2. Results will be saved in *.sm and *.smpvalue files
Created by: kah siong
Created on: 05/08/17
}
\author{
Kah Siong, Tan
}
